requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
lxml>=5.0.0
anthropic>=0.40.0
openai>=1.0.0
tenacity>=8.0.0
